@@ -1,21 +1,31 @@
 package rockhopper
 
 import (
-	"io/ioutil"
+	"os"
 
 	"github.com/codingconcepts/env"
 	"gopkg.in/yaml.v3"
 )
 
 type Config struct {
-	Driver        string `json:"driver" yaml:"driver" env:"ROCKHOPPER_DRIVER"`
-	Dialect       string `json:"dialect" yaml:"dialect" env:"ROCKHOPPER_DIALECT"`
-	DSN           string `json:"dsn" yaml:"dsn" env:"ROCKHOPPER_DSN"`
+	Driver  string `json:"driver" yaml:"driver" env:"ROCKHOPPER_DRIVER"`
+	Dialect string `json:"dialect" yaml:"dialect" env:"ROCKHOPPER_DIALECT"`
+	DSN     string `json:"dsn" yaml:"dsn" env:"ROCKHOPPER_DSN"`
+
+	Package string `json:"package" yaml:"package"`
+
 	MigrationsDir string `json:"migrationsDir" yaml:"migrationsDir" env:"ROCKHOPPER_MIGRATIONS_DIR"`
+
+	MigrationsDirs []string `json:"migrationsDirs" yaml:"migrationsDirs" env:"ROCKHOPPER_MIGRATIONS_DIRS"`
+
+	TableName string `json:"tableName" yaml:"tableName" env:"ROCKHOPPER_TABLE_NAME"`
+
+	// IncludePackages is used as a whitelist for the migration packages, optional
+	IncludePackages []string `json:"includePackages" yaml:"includePackages"`
 }
 
 func LoadConfig(configFile string) (*Config, error) {
-	data, err := ioutil.ReadFile(configFile)
+	data, err := os.ReadFile(configFile)
 	if err != nil {
 		return nil, err
 	}
@@ -27,6 +37,10 @@ func LoadConfig(configFile string) (*Config, error) {
 
 	if err := env.Set(&config); err != nil {
 		return nil, err
+	}
+
+	if len(config.MigrationsDir) > 0 {
+		config.MigrationsDirs = append(config.MigrationsDirs, config.MigrationsDir)
 	}
 
 	return &config, err
