@@ -2,6 +2,7 @@ package main
 
 import (
 	"context"
+	"errors"
 
 	"github.com/spf13/cobra"
 
@@ -23,27 +24,40 @@ var RedoCmd = &cobra.Command{
 }
 
 func redo(cmd *cobra.Command, args []string) error {
-	db, err := rockhopper.OpenByConfig(config)
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
+	db, err := rockhopper.OpenWithConfig(config)
 	if err != nil {
 		return err
 	}
 
 	defer db.Close()
 
-	loader := &rockhopper.SqlMigrationLoader{}
-	migrations, err := loader.Load(config.MigrationsDir)
+	if err := db.Touch(ctx); err != nil {
+		return err
+	}
+
+	loader := rockhopper.NewSqlMigrationLoader(config)
+
+	migrations, err := loader.Load(config.MigrationsDirs...)
 	if err != nil {
 		return err
 	}
 
-	currentVersion, err := db.CurrentVersion()
+	if len(migrations) == 0 {
+		log.Infof("no migrations found")
+		return nil
+	}
+
+	_, lastAppliedMigration, err := db.FindLastAppliedMigration(ctx, migrations)
 	if err != nil {
 		return err
 	}
 
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
+	if lastAppliedMigration == nil {
+		return errors.New("no migration has been applied yet")
+	}
 
-	return rockhopper.Redo(ctx, db, currentVersion, migrations)
+	return rockhopper.Redo(ctx, db, lastAppliedMigration)
 }
-
