@@ -3,11 +3,13 @@ package main
 import (
 	"context"
 
-	log "github.com/sirupsen/logrus"
+	"github.com/sirupsen/logrus"
 	"github.com/spf13/cobra"
 
 	"github.com/c9s/rockhopper"
 )
+
+var log = logrus.WithField("application", "rockhopper")
 
 func init() {
 	UpCmd.Flags().Int64("to", 0, "up to a specific version")
@@ -17,7 +19,7 @@ func init() {
 
 var UpCmd = &cobra.Command{
 	Use:   "up",
-	Short: "upgrade database",
+	Short: "run migration scripts to upgrade database schema",
 
 	// SilenceUsage is an option to silence usage when an error occurs.
 	SilenceUsage: true,
@@ -25,6 +27,9 @@ var UpCmd = &cobra.Command{
 }
 
 func up(cmd *cobra.Command, args []string) error {
+	ctx, cancel := context.WithCancel(context.Background())
+	defer cancel()
+
 	if err := checkConfig(config); err != nil {
 		return err
 	}
@@ -39,35 +44,69 @@ func up(cmd *cobra.Command, args []string) error {
 		return err
 	}
 
-	db, err := rockhopper.OpenByConfig(config)
+	db, err := rockhopper.OpenWithConfig(config)
 	if err != nil {
 		return err
 	}
 
 	defer db.Close()
 
-	loader := &rockhopper.SqlMigrationLoader{}
-	migrations, err := loader.Load(config.MigrationsDir)
+	if err := db.Touch(ctx); err != nil {
+		return err
+	}
+
+	loader := rockhopper.NewSqlMigrationLoader(config)
+
+	allMigrations, err := loader.Load(config.MigrationsDirs...)
 	if err != nil {
 		return err
 	}
 
-	currentVersion, err := db.CurrentVersion()
-	if err != nil {
-		return err
+	if len(allMigrations) == 0 {
+		log.Infof("no migrations found")
+		return nil
 	}
 
-	ctx, cancel := context.WithCancel(context.Background())
-	defer cancel()
+	debugMigrations(allMigrations)
 
-	if steps > 0 {
-		return rockhopper.UpBySteps(ctx, db, migrations, currentVersion, steps, func(m *rockhopper.Migration) {
-			log.Infof("migration %v is applied", m.Version)
-		})
+	migrationMap := allMigrations.MapByPackage()
+
+	if len(config.IncludePackages) > 0 {
+		migrationMap = migrationMap.FilterPackage(config.IncludePackages)
 	}
 
-	return rockhopper.Up(ctx, db, migrations, currentVersion, to, func(m *rockhopper.Migration) {
-		log.Infof("migration %v is applied", m.Version)
-	})
+	migrationMap = migrationMap.SortAndConnect()
 
+	for pkgName, migrations := range migrationMap {
+		_ = pkgName
+
+		_, lastAppliedMigration, err := db.FindLastAppliedMigration(ctx, migrations)
+		if err != nil {
+			return err
+		}
+
+		startMigration := migrations.Head()
+		if lastAppliedMigration != nil {
+			startMigration = lastAppliedMigration.Next
+		}
+
+		if steps > 0 {
+			err = rockhopper.UpBySteps(ctx, db, startMigration, steps, func(m *rockhopper.Migration) {
+				// log.Infof("migration %v is applied", m.Version)
+			})
+			if err != nil {
+				return err
+			}
+		} else {
+			err = rockhopper.Up(ctx, db, startMigration, to, func(m *rockhopper.Migration) {
+				// log.Infof("migration %d is applied", m.Version)
+			})
+			if err != nil {
+				return err
+			}
+		}
+
+	}
+
+	return nil
 }
