@@ -10,6 +10,17 @@ import (
 	log "github.com/sirupsen/logrus"
 )
 
+const (
+	VersionGoose        = 0
+	VersionRockhopperV1 = 1
+)
+
+// legacyGooseTableName is the legacy table name
+const legacyGooseTableName = "goose_db_version"
+
+// TableName is the migration version table name
+const TableName = "rockhopper_versions"
+
 type SQLExecutor interface {
 	ExecContext(ctx context.Context, query string, args ...interface{}) (sql.Result, error)
 }
@@ -22,7 +33,7 @@ type DB struct {
 	tableName  string
 }
 
-func OpenByConfig(config *Config) (*DB, error) {
+func OpenWithConfig(config *Config) (*DB, error) {
 	dialectName := config.Dialect
 	if len(dialectName) == 0 {
 		dialectName = config.Driver
@@ -41,7 +52,7 @@ func OpenByConfig(config *Config) (*DB, error) {
 		}
 	}
 
-	return Open(config.Driver, dialect, dsn)
+	return Open(config.Driver, dialect, dsn, TableName)
 }
 
 func BuildDSNFromEnvVars(driver string) (string, error) {
@@ -51,56 +62,6 @@ func BuildDSNFromEnvVars(driver string) (string, error) {
 
 	}
 	return "", fmt.Errorf("can not build dsn for driver %s", driver)
-}
-
-// buildMySqlDSN builds the data source name from environment variables
-func buildMySqlDSN() (string, error) {
-	if v, ok := os.LookupEnv("MYSQL_URL"); ok {
-		return v, nil
-	}
-
-	if v, ok := os.LookupEnv("MYSQL_DSN"); ok {
-		return v, nil
-	}
-
-	dsn := ""
-	user := "root"
-
-	if v, ok := os.LookupEnv("MYSQL_USER"); ok {
-		user = v
-		dsn += v
-	}
-
-	if v, ok := os.LookupEnv("MYSQL_PASSWORD"); ok {
-		dsn += ":" + v
-	} else if v, ok := os.LookupEnv("MYSQL_PASS"); ok {
-		dsn += ":" + v
-	} else if user == "root" {
-		if v, ok := os.LookupEnv("MYSQL_ROOT_PASSWORD"); ok {
-			dsn = ":" + v
-		}
-	}
-
-	address := ""
-	if v, ok := os.LookupEnv("MYSQL_HOST"); ok {
-		address = v
-	}
-
-	if v, ok := os.LookupEnv("MYSQL_PORT"); ok {
-		address += ":" + v
-	}
-
-	if v, ok := os.LookupEnv("MYSQL_PROTOCOL"); ok {
-		dsn += v + "(" + address + ")"
-	} else {
-		dsn += "tcp(" + address + ")"
-	}
-
-	if v, ok := os.LookupEnv("MYSQL_DATABASE"); ok {
-		dsn += "/" + v
-	}
-
-	return dsn, nil
 }
 
 func castDriverName(driver string) string {
@@ -116,17 +77,28 @@ func castDriverName(driver string) string {
 	return driver
 }
 
-func New(driverName string, dialect SQLDialect, db *sql.DB) *DB {
-	return &DB{
-		dialect:    dialect,
-		driverName: driverName,
-		DB:         db,
-		tableName:  defaultTableName,
+func OpenWithEnv(prefix string) (*DB, error) {
+	driverName := os.Getenv(prefix + "_DRIVER")
+	if driverName == "" {
+		return nil, fmt.Errorf("env %s_DRIVER is not defined", prefix)
 	}
+
+	dialectName := os.Getenv(prefix + "_DIALECT")
+	if dialectName == "" {
+		dialectName = driverName
+	}
+
+	dialect, err := LoadDialect(dialectName)
+	if err != nil {
+		return nil, err
+	}
+
+	dsn := os.Getenv(prefix + "_DSN")
+	return Open(driverName, dialect, dsn, TableName)
 }
 
 // Open creates a connection to a database
-func Open(driverName string, dialect SQLDialect, dsn string) (*DB, error) {
+func Open(driverName string, dialect SQLDialect, dsn string, tableName string) (*DB, error) {
 	driverName = castDriverName(driverName)
 
 	switch driverName {
@@ -141,45 +113,90 @@ func Open(driverName string, dialect SQLDialect, dsn string) (*DB, error) {
 		return nil, err
 	}
 
-	return New(driverName, dialect, db), nil
+	return New(driverName, dialect, db, tableName), nil
 }
 
-func (db *DB) deleteVersion(ctx context.Context, tx SQLExecutor, version int64) error {
-	if _, err := tx.ExecContext(ctx, db.dialect.deleteVersionSQL(db.tableName), version); err != nil {
+func New(driverName string, dialect SQLDialect, db *sql.DB, tableName string) *DB {
+	return &DB{
+		dialect:    dialect,
+		driverName: driverName,
+		DB:         db,
+		tableName:  tableName,
+	}
+}
+
+func (db *DB) deleteVersion(ctx context.Context, tx SQLExecutor, pkgName string, version int64) error {
+	if _, err := tx.ExecContext(ctx, db.dialect.deleteVersionSQL(db.tableName), pkgName, version); err != nil {
 		return errors.Wrap(err, "failed to delete migration record")
 	}
 
 	return nil
 }
 
-func (db *DB) insertVersion(ctx context.Context, tx SQLExecutor, version int64) error {
-	if _, err := tx.ExecContext(ctx, db.dialect.insertVersionSQL(db.tableName), version, true); err != nil {
+func (db *DB) getTableNames(ctx context.Context) ([]string, error) {
+	q := db.dialect.getTableNamesSQL()
+	rows, err := db.QueryContext(ctx, q)
+	if err != nil {
+		return nil, err
+	}
+
+	defer func() {
+		_ = rows.Close()
+	}()
+
+	var tableNames []string
+	for rows.Next() {
+		var tableName string
+		if err := rows.Scan(&tableName); err != nil {
+			return tableNames, err
+		}
+
+		tableNames = append(tableNames, tableName)
+	}
+
+	return tableNames, nil
+}
+
+func (db *DB) insertVersion(ctx context.Context, tx SQLExecutor, pkgName string, version int64, applied bool) error {
+	if _, err := tx.ExecContext(ctx, db.dialect.insertVersionSQL(db.tableName), pkgName, version, applied); err != nil {
 		return errors.Wrap(err, "failed to insert new migration record")
 	}
 
 	return nil
 }
 
-func (db *DB) FindMigration(version int64) (*MigrationRecord, error) {
-	var row MigrationRecord
+// LoadMigration finds the migration record from the db, and then updates the record to
+// the given Migration object. The migration.Record field will be updated.
+// When returning (nil, nil), it means the record is not found.
+func (db *DB) LoadMigration(ctx context.Context, m *Migration) (*Migration, error) {
+	var record MigrationRecord
 
 	var q = db.dialect.migrationSQL(db.tableName)
-	var err = db.QueryRow(q, version).Scan(&row.Time, &row.IsApplied)
 
-	if err != nil {
-		if err == sql.ErrNoRows {
-			return nil, nil
-		} else {
-			return nil, errors.Wrap(err, "failed to query the latest migration")
-		}
+	row := db.QueryRowContext(ctx, q, m.Package, m.Version)
+	if err := row.Err(); err != nil {
+		return nil, convertNoRowsErrToNil(err)
 	}
 
-	return &row, nil
+	var id int64
+	var err = row.Scan(&id, &record.Time, &record.IsApplied)
+	if err != nil {
+		return nil, convertNoRowsErrToNil(err)
+	}
 
+	m.Record = &record
+	return m, nil
 }
 
+// LoadMigrationRecords
+//
+// Deprecated: use LoadMigrationRecordsByPackage instead
 func (db *DB) LoadMigrationRecords() ([]MigrationRecord, error) {
-	rows, err := db.dialect.dbVersionQuery(db.DB, db.tableName)
+	return db.LoadMigrationRecordsByPackage(context.Background(), DefaultPackageName)
+}
+
+func (db *DB) LoadMigrationRecordsByPackage(ctx context.Context, pkgName string) ([]MigrationRecord, error) {
+	rows, err := db.DB.QueryContext(ctx, db.dialect.queryVersionsSQL(db.tableName), pkgName)
 	if err != nil {
 		return nil, err
 	}
@@ -197,7 +214,7 @@ func (db *DB) LoadMigrationRecords() ([]MigrationRecord, error) {
 	var records []MigrationRecord
 	for rows.Next() {
 		var row MigrationRecord
-		if err = rows.Scan(&row.VersionID, &row.IsApplied); err != nil {
+		if err = rows.Scan(&row.Package, &row.VersionID, &row.IsApplied, &row.Time); err != nil {
 			return nil, errors.Wrap(err, "failed to scan row")
 		}
 
@@ -211,80 +228,210 @@ func (db *DB) LoadMigrationRecords() ([]MigrationRecord, error) {
 	return records, nil
 }
 
-func (db *DB) CurrentVersion() (int64, error) {
-	rows, err := db.dialect.dbVersionQuery(db.DB, db.tableName)
+// runCoreMigration executes the core migration
+func (db *DB) runCoreMigration(ctx context.Context) error {
+	tableNames, err := db.getTableNames(ctx)
 	if err != nil {
-		// table exists, but there is no rows
-		if err == sql.ErrNoRows {
+		return err
+	}
+
+	// check if it's the latest version
+	if sliceContains(tableNames, TableName) {
+		// if so, we are good
+
+		// check the latest core version
+		latestVersion, err := db.queryLatestVersion(ctx, CorePackageName)
+		if err != nil {
+			return err
+		}
+
+		log.Debugf("found latest core package version: %d", latestVersion)
+
+		return db.upgradeCoreMigrations(ctx, latestVersion)
+	} else if sliceContains(tableNames, legacyGooseTableName) {
+
+		// the legacy version
+		log.Debugf("found legacy goose table, migrating...")
+
+		return db.migrateLegacyGooseTable(ctx)
+	}
+
+	// no version table found, create the version table with the latest schema
+	return db.createVersionTable(ctx, db, VersionRockhopperV1)
+}
+
+func (db *DB) upgradeCoreMigrations(ctx context.Context, currentVersion int64) error {
+	if currentVersion < 1 { /* do something */
+	}
+	if currentVersion < 2 { /* do something */
+	}
+	return nil
+}
+
+// queryLatestVersion selects the latest db version of a package
+func (db *DB) queryLatestVersion(ctx context.Context, pkgName string) (int64, error) {
+	row := db.DB.QueryRowContext(ctx,
+		db.dialect.selectLastVersionSQL(TableName), pkgName)
+
+	if err := row.Err(); err != nil {
+		return 0, convertNoRowsErrToNil(err)
+	}
+
+	var versionId sql.NullInt64
+	if err := row.Scan(&versionId); err != nil {
+		return 0, err
+	}
+
+	if versionId.Valid {
+		return versionId.Int64, nil
+	}
+
+	return 0, nil
+}
+
+// migrateLegacyGooseTable migrates the legacy goose version table to the new rockhopper version table
+func (db *DB) migrateLegacyGooseTable(ctx context.Context) error {
+	if err := db.createVersionTable(ctx, db, VersionRockhopperV1); err != nil {
+		return err
+	}
+
+	tx, err := db.DB.Begin()
+	if err != nil {
+		return err
+	}
+
+	switch db.dialect.(type) {
+	case *MySQLDialect, *TiDBDialect:
+		if err := execAndCheckErr(tx, ctx,
+			`ALTER TABLE goose_db_version ADD COLUMN package VARCHAR(125) NOT NULL DEFAULT 'main'`); err != nil {
+			return rollbackAndLogErr(err, tx, "unable to alter table")
+		}
+
+		if err := execAndCheckErr(tx, ctx,
+			fmt.Sprintf(`RENAME TABLE %s TO %s`, legacyGooseTableName, TableName)); err != nil {
+			return rollbackAndLogErr(err, tx, "unable to rename table")
+		}
+
+	case *PostgresDialect, *RedshiftDialect:
+		if err := execAndCheckErr(tx, ctx,
+			`ALTER TABLE goose_db_version ADD COLUMN package VARCHAR(125) NOT NULL DEFAULT 'main'`); err != nil {
+			return rollbackAndLogErr(err, tx, "unable to alter table")
+		}
+
+		if err := execAndCheckErr(tx, ctx,
+			fmt.Sprintf(`ALTER TABLE %s RENAME TO %s`, legacyGooseTableName, TableName)); err != nil {
+			return rollbackAndLogErr(err, tx, "unable to rename table")
+		}
+
+	case *Sqlite3Dialect, *SqlServerDialect:
+		if err := execAndCheckErr(tx, ctx,
+			fmt.Sprintf(`INSERT INTO %s(id, package, version_id, is_applied, tstamp) SELECT id, 'main', version_id, is_applied, tstamp FROM %s`,
+				TableName,
+				legacyGooseTableName),
+		); err != nil {
+			return rollbackAndLogErr(err, tx, "unable to execute insert from select")
+		}
+
+		if err := execAndCheckErr(tx, ctx, fmt.Sprintf(`DROP TABLE %s`, legacyGooseTableName)); err != nil {
+			return rollbackAndLogErr(err, tx, "unable to drop legacy table")
+		}
+	}
+
+	return tx.Commit()
+}
+
+// Touch checks if the version table exists, if not, create the version table
+func (db *DB) Touch(ctx context.Context) error {
+	if err := db.runCoreMigration(ctx); err != nil {
+		return err
+	}
+
+	_, err := db.queryLatestVersion(ctx, CorePackageName)
+	return convertNoRowsErrToNil(err)
+}
+
+// CurrentVersion get the current version of the migration version table
+func (db *DB) CurrentVersion(ctx context.Context, packageName string) (int64, error) {
+	if err := db.Touch(ctx); err != nil {
+		return 0, err
+	}
+
+	version, err := db.queryLatestVersion(ctx, packageName)
+	if err != nil {
+		if errors.Is(err, sql.ErrNoRows) {
 			return 0, nil
 		}
 
-		return 0, db.createVersionTable()
-	}
-
-	if err := rows.Close(); err != nil {
 		return 0, err
 	}
 
-	records, err := db.LoadMigrationRecords()
-	if err != nil {
-		return 0, err
-	}
-
-	// The most recent record for each migration specifies
-	// whether it has been applied or rolled back.
-	// The first version we find that has been applied is the current version.
-	toSkip := make([]int64, 0)
-	for _, row := range records {
-		// have we already marked this version to be skipped?
-		skip := false
-		for _, v := range toSkip {
-			if v == row.VersionID {
-				skip = true
-				break
-			}
-		}
-
-		if skip {
-			continue
-		}
-
-		// if version has been applied we're done
-		if row.IsApplied {
-			return row.VersionID, nil
-		}
-
-		// latest version of migration has not been applied.
-		toSkip = append(toSkip, row.VersionID)
-	}
-
-	return 0, ErrNoCurrentVersion
+	return version, nil
 }
 
-// Create the db version table
-// and insert the initial 0 value into it
-func (db *DB) createVersionTable() error {
-	txn, err := db.Begin()
+// createVersionTable creates the db version table and inserts the initial value 0 into the migration table
+func (db *DB) createVersionTable(ctx context.Context, tx SqlExecutor, initVersion int64) error {
+	if _, err := tx.ExecContext(ctx, db.dialect.createVersionTableSQL(db.tableName)); err != nil {
+		return err
+	}
+
+	return db.insertVersion(ctx, tx, CorePackageName, initVersion, true)
+}
+
+func sliceContains(a []string, b string) bool {
+	for _, s := range a {
+		if s == b {
+			return true
+		}
+	}
+	return false
+}
+
+func rollbackAndLogErr(originErr error, txn *sql.Tx, msg string, args ...any) error {
+	if err := txn.Rollback(); err != nil {
+		log.WithError(err).Errorf("unable to rollback transaction")
+		if len(msg) > 0 {
+			return errors.Wrapf(originErr, msg, args...)
+		}
+
+		return errors.Wrapf(originErr, err.Error())
+	}
+
+	return originErr
+}
+
+func execAndCheckErr(db SqlExecutor, ctx context.Context, sql string, args ...interface{}) error {
+	_, err := db.ExecContext(ctx, sql, args...)
 	if err != nil {
+		log.WithError(err).Errorf("unable to execute SQL: %s", sql)
 		return err
 	}
 
-	if _, err := txn.Exec(db.dialect.createVersionTableSQL(db.tableName)); err != nil {
-		if err := txn.Rollback(); err != nil {
-			log.WithError(err).Error("create version table, rollback error")
+	return nil
+}
+
+func convertNoRowsErrToNil(err error) error {
+	if errors.Is(err, sql.ErrNoRows) {
+		return nil
+	}
+
+	return err
+}
+
+func (db *DB) FindLastAppliedMigration(
+	ctx context.Context, allMigrations MigrationSlice,
+) (int, *Migration, error) {
+	for i := len(allMigrations) - 1; i >= 0; i-- {
+		m := allMigrations[i]
+
+		m, err := db.LoadMigration(ctx, m)
+		if err != nil {
+			return -1, nil, err
 		}
-		return err
-	}
 
-	version := 0
-	applied := true
-	if _, err := txn.Exec(db.dialect.insertVersionSQL(db.tableName), version, applied); err != nil {
-		if err := txn.Rollback(); err != nil {
-			log.WithError(err).Error("insert version, rollback error")
+		if m != nil && m.Record != nil && m.Record.IsApplied {
+			return i, m, nil
 		}
-
-		return err
 	}
 
-	return txn.Commit()
+	return -1, nil, nil
 }
