@@ -5,20 +5,22 @@ import (
 	"fmt"
 )
 
-var defaultTableName = "goose_db_version"
-
-// TableName returns goose db version table name
-func TableName() string {
-	return defaultTableName
-}
-
 // SQLDialect abstracts the details of specific SQL dialects
 // for goose's few SQL specific statements
 type SQLDialect interface {
+	getTableNamesSQL() string                      // return the sql string to get the table names
 	createVersionTableSQL(tableName string) string // sql string to create the db version table
 	insertVersionSQL(tableName string) string      // sql string to insert the initial version table row
 	deleteVersionSQL(tableName string) string      // sql string to delete version
-	migrationSQL(tableName string) string          // sql string to retrieve migrations
+
+	// migrationSQL returns the sql string to retrieve migrations
+	migrationSQL(tableName string) string
+
+	// selectLastVersionSQL returns the sql string to get the latest version
+	selectLastVersionSQL(tableName string) string
+
+	// queryVersionsSQL returns the sql string to query the version info descending
+	queryVersionsSQL(tableName string) string
 	dbVersionQuery(db *sql.DB, tableName string) (*sql.Rows, error)
 }
 
@@ -39,80 +41,4 @@ func LoadDialect(d string) (SQLDialect, error) {
 	}
 
 	return nil, fmt.Errorf("%q: unknown dialect", d)
-}
-
-////////////////////////////
-// Redshift
-////////////////////////////
-
-// RedshiftDialect struct.
-type RedshiftDialect struct{}
-
-func (d RedshiftDialect) createVersionTableSQL(tableName string) string {
-	return fmt.Sprintf(`CREATE TABLE %s (
-            	id integer NOT NULL identity(1, 1),
-                version_id bigint NOT NULL,
-                is_applied boolean NOT NULL,
-                tstamp timestamp NULL default sysdate,
-                PRIMARY KEY(id)
-            );`, tableName)
-}
-
-func (d RedshiftDialect) insertVersionSQL(tableName string) string {
-	return fmt.Sprintf("INSERT INTO %s (version_id, is_applied) VALUES ($1, $2);", tableName)
-}
-
-func (d RedshiftDialect) dbVersionQuery(db *sql.DB, tableName string) (*sql.Rows, error) {
-	rows, err := db.Query(fmt.Sprintf("SELECT version_id, is_applied from %s ORDER BY id DESC", tableName))
-	if err != nil {
-		return nil, err
-	}
-
-	return rows, err
-}
-
-func (d RedshiftDialect) migrationSQL(tableName string) string {
-	return fmt.Sprintf("SELECT tstamp, is_applied FROM %s WHERE version_id=$1 ORDER BY tstamp DESC LIMIT 1", tableName)
-}
-
-func (d RedshiftDialect) deleteVersionSQL(tableName string) string {
-	return fmt.Sprintf("DELETE FROM %s WHERE version_id=$1;", tableName)
-}
-
-////////////////////////////
-// TiDB
-////////////////////////////
-
-// TiDBDialect struct.
-type TiDBDialect struct{}
-
-func (m TiDBDialect) createVersionTableSQL(tableName string) string {
-	return fmt.Sprintf(`CREATE TABLE %s (
-                id BIGINT UNSIGNED NOT NULL AUTO_INCREMENT UNIQUE,
-                version_id bigint NOT NULL,
-                is_applied boolean NOT NULL,
-                tstamp timestamp NULL default now(),
-                PRIMARY KEY(id)
-            );`, tableName)
-}
-
-func (m TiDBDialect) insertVersionSQL(tableName string) string {
-	return fmt.Sprintf("INSERT INTO %s (version_id, is_applied) VALUES (?, ?);", tableName)
-}
-
-func (m TiDBDialect) dbVersionQuery(db *sql.DB, tableName string) (*sql.Rows, error) {
-	rows, err := db.Query(fmt.Sprintf("SELECT version_id, is_applied from %s ORDER BY id DESC", tableName))
-	if err != nil {
-		return nil, err
-	}
-
-	return rows, err
-}
-
-func (m TiDBDialect) migrationSQL(tableName string) string {
-	return fmt.Sprintf("SELECT tstamp, is_applied FROM %s WHERE version_id=? ORDER BY tstamp DESC LIMIT 1", tableName)
-}
-
-func (m TiDBDialect) deleteVersionSQL(tableName string) string {
-	return fmt.Sprintf("DELETE FROM %s WHERE version_id=?;", tableName)
 }
