@@ -8,6 +8,10 @@ import (
 // SqlServerDialect struct.
 type SqlServerDialect struct{}
 
+func (m SqlServerDialect) getTableNamesSQL() string {
+	return `SELECT * FROM SYS.TABLES;`
+}
+
 func (m SqlServerDialect) createVersionTableSQL(tableName string) string {
 	return fmt.Sprintf(`CREATE TABLE %s (
                 id INT NOT NULL IDENTITY(1,1) PRIMARY KEY,
@@ -18,11 +22,19 @@ func (m SqlServerDialect) createVersionTableSQL(tableName string) string {
 }
 
 func (m SqlServerDialect) insertVersionSQL(tableName string) string {
-	return fmt.Sprintf("INSERT INTO %s (version_id, is_applied) VALUES (@p1, @p2);", tableName)
+	return fmt.Sprintf("INSERT INTO %s (package, version_id, is_applied) VALUES (@p1, @p2, @p3);", tableName)
+}
+
+func (m SqlServerDialect) selectLastVersionSQL(tableName string) string {
+	return fmt.Sprintf("SELECT MAX(version_id) FROM %s WHERE package = @p1", tableName)
+}
+
+func (m SqlServerDialect) queryVersionsSQL(tableName string) string {
+	return fmt.Sprintf("SELECT package, version_id, is_applied, tstamp FROM %s WHERE package = @p1 ORDER BY id DESC", tableName)
 }
 
 func (m SqlServerDialect) dbVersionQuery(db *sql.DB, tableName string) (*sql.Rows, error) {
-	rows, err := db.Query(fmt.Sprintf("SELECT version_id, is_applied FROM %s ORDER BY id DESC", tableName))
+	rows, err := db.Query(fmt.Sprintf("SELECT package, version_id, is_applied FROM %s ORDER BY id DESC", tableName))
 	if err != nil {
 		return nil, err
 	}
@@ -34,12 +46,12 @@ func (m SqlServerDialect) migrationSQL(tableName string) string {
 	const tpl = `
 WITH Migrations AS
 (
-    SELECT tstamp, is_applied,
+    SELECT id, tstamp, is_applied,
     ROW_NUMBER() OVER (ORDER BY tstamp) AS 'RowNumber'
     FROM %s 
-	WHERE version_id=@p1
+	WHERE package = @p1 AND version_id = @p2
 ) 
-SELECT tstamp, is_applied 
+SELECT id, tstamp, is_applied 
 FROM Migrations 
 WHERE RowNumber BETWEEN 1 AND 2
 ORDER BY tstamp DESC
@@ -48,5 +60,5 @@ ORDER BY tstamp DESC
 }
 
 func (m SqlServerDialect) deleteVersionSQL(tableName string) string {
-	return fmt.Sprintf("DELETE FROM %s WHERE version_id = @p1", tableName)
+	return fmt.Sprintf("DELETE FROM %s WHERE package = @p1 AND version_id = @p2", tableName)
 }
