@@ -8,22 +8,36 @@ import (
 // MySQLDialect struct.
 type MySQLDialect struct{}
 
+func (m MySQLDialect) getTableNamesSQL() string {
+	return `SHOW TABLES`
+}
+
 func (m MySQLDialect) createVersionTableSQL(tableName string) string {
 	return fmt.Sprintf(`CREATE TABLE IF NOT EXISTS %s (
-                id serial NOT NULL,
-                version_id bigint NOT NULL,
-                is_applied boolean NOT NULL,
-                tstamp timestamp NULL default now(),
-                PRIMARY KEY(id)
+                id SERIAL NOT NULL,
+                package VARCHAR(125) NOT NULL DEFAULT 'main',
+                version_id BIGINT NOT NULL,
+                is_applied BOOLEAN NOT NULL,
+                tstamp TIMESTAMP NOT NULL DEFAULT NOW(),
+                PRIMARY KEY(id),
+				UNIQUE unique_version(version_id)
             );`, tableName)
 }
 
 func (m MySQLDialect) insertVersionSQL(tableName string) string {
-	return fmt.Sprintf("INSERT INTO %s (version_id, is_applied) VALUES (?, ?);", tableName)
+	return fmt.Sprintf("INSERT INTO %s (package, version_id, is_applied) VALUES (?, ?, ?);", tableName)
+}
+
+func (m MySQLDialect) selectLastVersionSQL(tableName string) string {
+	return fmt.Sprintf("SELECT MAX(version_id) FROM %s WHERE package = ?", tableName)
+}
+
+func (m MySQLDialect) queryVersionsSQL(tableName string) string {
+	return fmt.Sprintf("SELECT package, version_id, is_applied, tstamp FROM %s WHERE package = ? ORDER BY id DESC", tableName)
 }
 
 func (m MySQLDialect) dbVersionQuery(db *sql.DB, tableName string) (*sql.Rows, error) {
-	rows, err := db.Query(fmt.Sprintf("SELECT version_id, is_applied FROM %s ORDER BY id DESC", tableName))
+	rows, err := db.Query(fmt.Sprintf("SELECT package, version_id, is_applied FROM %s ORDER BY id DESC", tableName))
 	if err != nil {
 		return nil, err
 	}
@@ -32,9 +46,9 @@ func (m MySQLDialect) dbVersionQuery(db *sql.DB, tableName string) (*sql.Rows, e
 }
 
 func (m MySQLDialect) migrationSQL(tableName string) string {
-	return fmt.Sprintf("SELECT tstamp, is_applied FROM %s WHERE version_id=? ORDER BY tstamp DESC LIMIT 1", tableName)
+	return fmt.Sprintf("SELECT id, tstamp, is_applied FROM %s WHERE package = ? AND version_id = ? ORDER BY tstamp DESC LIMIT 1", tableName)
 }
 
 func (m MySQLDialect) deleteVersionSQL(tableName string) string {
-	return fmt.Sprintf("DELETE FROM %s WHERE version_id = ?", tableName)
+	return fmt.Sprintf("DELETE FROM %s WHERE package = ? AND version_id = ?", tableName)
 }
