@@ -8,9 +8,14 @@ import (
 // Sqlite3Dialect struct.
 type Sqlite3Dialect struct{}
 
+func (m Sqlite3Dialect) getTableNamesSQL() string {
+	return `SELECT name FROM sqlite_master WHERE type='table'`
+}
+
 func (m Sqlite3Dialect) createVersionTableSQL(tableName string) string {
 	return fmt.Sprintf(`CREATE TABLE IF NOT EXISTS %s (
                 id INTEGER PRIMARY KEY AUTOINCREMENT,
+                package TEXT NOT NULL DEFAULT 'main',
                 version_id INTEGER NOT NULL,
                 is_applied INTEGER NOT NULL,
                 tstamp TIMESTAMP DEFAULT (datetime('now'))
@@ -18,11 +23,19 @@ func (m Sqlite3Dialect) createVersionTableSQL(tableName string) string {
 }
 
 func (m Sqlite3Dialect) insertVersionSQL(tableName string) string {
-	return fmt.Sprintf("INSERT INTO %s (version_id, is_applied) VALUES (?, ?);", tableName)
+	return fmt.Sprintf("INSERT INTO %s (package, version_id, is_applied) VALUES (?, ?, ?);", tableName)
+}
+
+func (m Sqlite3Dialect) selectLastVersionSQL(tableName string) string {
+	return fmt.Sprintf("SELECT MAX(version_id) FROM %s WHERE package = ?", tableName)
+}
+
+func (m Sqlite3Dialect) queryVersionsSQL(tableName string) string {
+	return fmt.Sprintf("SELECT package, version_id, is_applied, tstamp FROM %s WHERE package = ? ORDER BY id DESC", tableName)
 }
 
 func (m Sqlite3Dialect) dbVersionQuery(db *sql.DB, tableName string) (*sql.Rows, error) {
-	rows, err := db.Query(fmt.Sprintf("SELECT version_id, is_applied from %s ORDER BY id DESC", tableName))
+	rows, err := db.Query(fmt.Sprintf("SELECT package, version_id, is_applied from %s ORDER BY id DESC", tableName))
 	if err != nil {
 		return nil, err
 	}
@@ -31,9 +44,9 @@ func (m Sqlite3Dialect) dbVersionQuery(db *sql.DB, tableName string) (*sql.Rows,
 }
 
 func (m Sqlite3Dialect) migrationSQL(tableName string) string {
-	return fmt.Sprintf("SELECT tstamp, is_applied FROM %s WHERE version_id=? ORDER BY tstamp DESC LIMIT 1", tableName)
+	return fmt.Sprintf("SELECT tstamp, is_applied FROM %s WHERE package = ? AND version_id = ? ORDER BY tstamp DESC LIMIT 1", tableName)
 }
 
 func (m Sqlite3Dialect) deleteVersionSQL(tableName string) string {
-	return fmt.Sprintf("DELETE FROM %s WHERE version_id = ?", tableName)
+	return fmt.Sprintf("DELETE FROM %s WHERE package = ? AND version_id = ?", tableName)
 }
