@@ -2,16 +2,7 @@ package rockhopper
 
 import "context"
 
-func DownBySteps(ctx context.Context, db *DB, migrations MigrationSlice, from int64, steps int, callbacks ...func(m *Migration)) error {
-	if len(migrations) == 0 {
-		return nil
-	}
-
-	m, err := migrations.Find(from)
-	if err != nil { // if no migration is found, we should not continue
-		return err
-	}
-
+func DownBySteps(ctx context.Context, db *DB, m *Migration, steps int, callbacks ...func(m *Migration)) error {
 	for ; steps > 0; steps-- {
 		if err := m.Down(ctx, db); err != nil {
 			return err
@@ -31,20 +22,13 @@ func DownBySteps(ctx context.Context, db *DB, migrations MigrationSlice, from in
 	return nil
 }
 
-func Down(ctx context.Context, db *DB, migrations MigrationSlice, from, to int64, callbacks ...func(m *Migration)) error {
-	if len(migrations) == 0 {
-		return nil
-	}
-
-	m, err := migrations.Find(from)
-	if err != nil { // if no migration is found, we should not continue
-		return err
-	}
-
-	for ; m.Previous != nil; m = m.Previous {
+func Down(ctx context.Context, db *DB, m *Migration, to int64, callbacks ...func(m *Migration)) error {
+	for ; m != nil; m = m.Previous {
 		if to > 0 && m.Version < to {
 			break
 		}
+
+		descMigration("downgrading", m)
 
 		if err := m.Down(ctx, db); err != nil {
 			return err
